Extract parse logical AND/Or and NOT expression to another file
